requirements.txt
django==4.1.2
djangorestframework
django-filter
drf-yasg
django-cors-headers
gunicorn
pymysql
django-ses
django-redis
django-storages
djangorestframework-simplejwt
django-resized
apns2
pillow